test: add more axelar unit tests
